feat: move key-up detection from keyboard to tui
Previously, keyboard handling was done considering only the KeyPress event.
The keys would 'tick' a certain duration after being pressed, and then reset.
However, the missing KeyUp events are specific to Bubbletea (and will be
resolved in the future), so they should be in the TUI rather than in the
keyboard.

This will make it easier to add an alternative GUI, as typically KeyUp
events *are* supported.