feat: integrated ebiten gui in the main.go app
In order to provide a graphical user interface (GUI) for the application,
a new package "gui" is introduced using the Ebiten 2D game library. This
facilitates a more interactive and visually appealing way of interacting
with the chip8.

Moreover, the main structure, the App, was modified to support multiple
user interfaces in addition to the existing Terminal User Interface
(TUI). It includes a switch statement to choose between the TUI and GUI
based on the argument passed via the command line.

New dependencies were also added to the go.mod and go.sum files to
support this feature.