feat: improve user interface selection and logging
Create 'gui.go' and 'tui.go' under the 'cmd' directory to separate the
implementations of the two User Interface types: GUI (Graphical User
Interface) and TUI (Text User Interface). This change allows easier
extension in future in case more interfaces are to be supported. The
registration of different interface types is taken care of in their
respective 'init()' functions, which are being added in the respective
files.

Using build tags, only include 'gui.go' in case we're on ARCH=amd64
and OS=windows/linux, as e.g. on Android this UI can't be build.

Modify 'main.go' to manage command-line flag parsing efficiently and
add documentation for easier understanding. An "App" interface and an
"AppBuilder" type are added. "UIs" type is also added, which is a map
of interface type's name and its builder function. This makes the
addition of new interfaces and their selection much simpler and
cleaner. Consequently, the logging system is updated to handle these
changes and it now logs the selected UI type as well. A help flag is
added as well that displays the usage of each flag. All these changes
have resulted in cleaner, modular and more maintainable code.

Add "-x" option debug flag in 'build.sh' to print each command before
execution, useful for tracing and monitoring the application build
process.